+++ src/inputs/multiple.js
@@ -121,6 +121,19 @@ var callSuper = Selectivity.inherits(MultipleInput, Selectivity, {
      * @inherit
      */
     filterResults: function(results) {
+        var hasChildren = results.some(function(item) {
+            return item.children;
+        });
+        if (hasChildren) {
+            results = results.map(function(item) {
+                return {
+                    id: item.id,
+                    text: item.text,
+                    children: this.filterResults(item.children)
+                };
+            }, this);
+        }
+
         return results.filter(function(item) {
             return !Selectivity.findById(this._data, item.id);
         }, this);